# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ONCF-2255-FO ExpressJS API
#during dev, should point to your local machine
#host: pprdweb-2255:8443
#host: localhost:8443
host: oncf-2255-api-dev.azurewebsites.net
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >
      For accessing the API a valid JWT token must be passed in all the queries
      in

      the 'Authorization' header.



      A valid JWT token is generated by the API and retourned as answer of a
      call

      to the route /login giving a valid user & password.



      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /Account/ActivateAccount:
    x-swagger-router-controller: account
    post:
      tags:
      - "Account"
      operationId: activateAccount
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: body
        name: body
        required: true
        schema:
         $ref: "#/definitions/activateAccountUserModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /Account/UnlockAccount:
    x-swagger-router-controller: account
    post:
      tags:
      - "Account"
      operationId: unlockAccount
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: body
        name: body
        required: true
        schema:
         $ref: "#/definitions/activateAccountUserModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /Account/ChangePassword:
    x-swagger-router-controller: account
    post:
      tags:
      - "Account"
      operationId: accountChangePassword
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: body
        name: body
        description: provide email, oldPassword and newPassword.
        required: true
        schema:
          $ref: "#/definitions/changePasswordModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /Account/ConfirmForgottenPassword:
    x-swagger-router-controller: account
    post:
      tags:
      - "Account"
      operationId: confirmForgottenPassword
      parameters:
      - in: body
        name: body
        description: provide confirmationCode, email and newPassword.
        required: true
        schema:
          $ref: "#/definitions/confirmForgottenPasswordmodel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /Account/ClientRegistration:
    x-swagger-router-controller: account
    post:
      tags:
      - "Account"
      operationId: clientRegistration
      parameters:
      - in: body
        name: body
        required: true
        schema:
         $ref: "#/definitions/clientRegistrationModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /MyProfile:
    x-swagger-router-controller: account
    get:
      tags:
      - "Account"
      operationId: myProfile
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /ForgotPassword:
    x-swagger-router-controller: account
    get:
      tags:
      - "Account"
      operationId: forgotPassword
      parameters:
      - name: email
        in: query
        required: true
        type: string
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ONCFResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
definitions:
  ONCFResponse:
    required:
      - data
      - error
      - isSuccess
    properties:
      data:
        type: object
      error:
        type: object
      isSuccess:
        type: boolean
  ErrorResponse:
    required:
      - data
      - error
      - isSuccess
    properties:
      data:
        type: object
      error:
        type: object
      isSuccess:
        type: boolean
...
  activateAccountUserModel:
    required:
      - email
      - code
    properties:
      email:
        type: string
      code:
        type: string
  changePasswordModel:
    required:
      - email
      - oldPassword
      - newPassword
    properties:
      email:
        type: string
      oldPassword:
        type: string
      newPassword:
        type: string
  confirmForgottenPasswordmodel:
    required:
      - confirmationCode
      - email
      - newPassword
    properties:
      confirmationCode:
        type: string
      email:
        type: string
      newPassword:
        type: string
  clientRegistrationModel:
    properties:
      title:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      mobileNumber:
        type: string
      city:
        type: string
      email:
        type: string
      password:
        type: string
...